Resolve multiple template error
Added a break inside the files.present? check to prevent the VC
templates from exceeding 1. This will potentially have an issue in the
case that there are more than one templates associated with a given
ViewComponent.rb file (i.e. erb + html) Pending test created with a
appeal for decision.

Co-authored-by: Eric Wanchic <ewanchic@g2.com>
Co-authored-by: Jesus Quezada-Guillen <jquezadaguillen@g2.com>
Co-authored-by: John Polhill <jpolhill@g2.com>